Learn the window increment not the window, half way through

+++ src/unicorn.hh
@@ -25,7 +25,7 @@ protected:
 
   long unsigned int _packets_sent, _packets_received;
   double _last_send_time;
-  int _the_window;
+  double _the_window;
   double _intersend_time;
 
   unsigned int _flow_id;
